woohoo the lib compiles

+++ include/anmalloc/anmalloc.h
@@ -12,7 +12,7 @@
 // - anmalloc_lock_t
 
 void anmalloc_free(void * buf);
-void * anmalloc_alloc(size_t size);
-void * anmalloc_aligned(size_t alignment, size_t size);
-void * anmalloc_realloc(void * ptr, size_t size);
+void * anmalloc_alloc(uint64_t size);
+void * anmalloc_aligned(uint64_t alignment, uint64_t size);
+void * anmalloc_realloc(void * ptr, uint64_t size);
